upd: library block acl

+++ lib/athena/content/library.ex
@@ -38,28 +38,41 @@ defmodule Athena.Content.Library do
     end
   end
 
-  @doc "Creates a new library block template."
-  @spec create_library_block(map()) :: {:ok, LibraryBlock.t()} | {:error, Ecto.Changeset.t()}
-  def create_library_block(attrs) do
-    %LibraryBlock{}
-    |> LibraryBlock.changeset(attrs)
-    |> Repo.insert()
+  @doc "Creates a new library block template. Sets owner to current user."
+  @spec create_library_block(map(), map()) ::
+          {:ok, LibraryBlock.t()} | {:error, Ecto.Changeset.t() | :unauthorized}
+  def create_library_block(user, attrs) do
+    if Identity.can?(user, "library.update") do
+      %LibraryBlock{owner_id: user.id}
+      |> LibraryBlock.changeset(attrs)
+      |> Repo.insert()
+    else
+      {:error, :unauthorized}
+    end
   end
 
-  @doc "Updates a library block template."
-  @spec update_library_block(LibraryBlock.t(), map()) ::
-          {:ok, LibraryBlock.t()} | {:error, Ecto.Changeset.t()}
-  def update_library_block(%LibraryBlock{} = block, attrs) do
-    block
-    |> LibraryBlock.changeset(attrs)
-    |> Repo.update()
+  @doc "Updates a library block template. Checks own_only policies."
+  @spec update_library_block(map(), LibraryBlock.t(), map()) ::
+          {:ok, LibraryBlock.t()} | {:error, Ecto.Changeset.t() | :unauthorized}
+  def update_library_block(user, %LibraryBlock{} = block, attrs) do
+    if Identity.can?(user, "library.update", block) do
+      block
+      |> LibraryBlock.changeset(attrs)
+      |> Repo.update()
+    else
+      {:error, :unauthorized}
+    end
   end
 
-  @doc "Deletes a library block template."
-  @spec delete_library_block(LibraryBlock.t()) ::
-          {:ok, LibraryBlock.t()} | {:error, Ecto.Changeset.t()}
-  def delete_library_block(%LibraryBlock{} = block) do
-    Repo.delete(block)
+  @doc "Deletes a library block template. Checks own_only policies."
+  @spec delete_library_block(map(), LibraryBlock.t()) ::
+          {:ok, LibraryBlock.t()} | {:error, Ecto.Changeset.t() | :unauthorized}
+  def delete_library_block(user, %LibraryBlock{} = block) do
+    if Identity.can?(user, "library.update", block) do
+      Repo.delete(block)
+    else
+      {:error, :unauthorized}
+    end
   end
 
   @doc """
